# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Audd Flow — AI-governed smart treasury & automated payroll on BOT Chain
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TokenInfo {
  symbol: string;
  address: string;
  decimals: number;
}

export interface NetworkConfig {
  name: string;
  chainId: number;
  rpcUrl: string;
  explorerUrl: string;
  faucetUrl: string;
  nativeToken: TokenInfo;
  ausdToken?: TokenInfo | null;
}

export interface DeploymentInfo {
  deployed: boolean;
  owner?: string | null;
  agent?: string | null;
  treasuryAddress?: string | null;
  ausdAddress?: string | null;
  deployedAtBlock?: number | null;
  deployedAt?: string | null;
}

export interface AppConfig {
  network: NetworkConfig;
  deployment: DeploymentInfo;
}

export interface TokenBalance {
  token: string;
  symbol: string;
  decimals: number;
  raw: string;
  formatted: string;
}

export interface Guardrail {
  token: string;
  symbol: string;
  perTxCap: string;
  perTxCapFormatted: string;
  dailyCap: string;
  dailyCapFormatted: string;
  spentToday: string;
  spentTodayFormatted: string;
  remainingToday: string;
  remainingTodayFormatted: string;
}

export interface TreasurySummary {
  treasuryAddress: string;
  paused: boolean;
  balances: TokenBalance[];
  guardrails: Guardrail[];
  employeeCount: number;
  activeEmployeeCount: number;
  monthlyPayroll: string;
  monthlyPayrollFormatted: string;
}

export interface Employee {
  address: string;
  name: string;
  role: string;
  salary: string;
  salaryFormatted: string;
  salaryToken: string;
  salarySymbol: string;
  active: boolean;
  addedAt: number;
}

export interface CreateEmployeeRequest {
  address: string;
  name: string;
  role: string;
  /** Human amount, e.g. '6000' */
  salary: string;
  /** 'aUSD' or 'BOT' */
  token?: string;
}

export interface UpdateEmployeeRequest {
  active: boolean;
}

export interface EmployeeMutationResult {
  txHash: string;
  explorerUrl: string;
  employee?: Employee;
}

export interface Stream {
  id: number;
  to: string;
  toName: string;
  token: string;
  symbol: string;
  ratePerSecond: string;
  ratePerDayFormatted: string;
  start: number;
  stop: number;
  lastClaim: number;
  claimed: string;
  claimedFormatted: string;
  claimable: string;
  claimableFormatted: string;
  active: boolean;
}

export interface CreateStreamRequest {
  to: string;
  /** Human amount per day, e.g. '200' */
  ratePerDay: string;
  token?: string;
  durationDays?: number | null;
}

export interface GuardrailUpdateRequest {
  /** 'aUSD' or 'BOT' */
  token: string;
  perTxCap: string;
  dailyCap: string;
}

export interface GuardrailCheck {
  ok: boolean;
  violations: string[];
}

export type PlanActionType = typeof PlanActionType[keyof typeof PlanActionType];


export const PlanActionType = {
  pay: 'pay',
  payroll: 'payroll',
  stream: 'stream',
  pause: 'pause',
  unpause: 'unpause',
  guardrail: 'guardrail',
  unknown: 'unknown',
} as const;

export interface PlanAction {
  type: PlanActionType;
  description: string;
  token?: string | null;
  tokenSymbol?: string | null;
  amount?: string | null;
  amountFormatted?: string | null;
  recipient?: string | null;
  recipientName?: string | null;
}

export interface AgentPlan {
  intent: string;
  category: string;
  feasible: boolean;
  notes?: string | null;
  actions: PlanAction[];
}

export type AgentIntentStatus = typeof AgentIntentStatus[keyof typeof AgentIntentStatus];


export const AgentIntentStatus = {
  planned: 'planned',
  executed: 'executed',
  failed: 'failed',
  rejected: 'rejected',
} as const;

export interface AgentIntent {
  id: number;
  prompt: string;
  category: string;
  summary: string;
  status: AgentIntentStatus;
  reasoning: string;
  plan: AgentPlan;
  validation: GuardrailCheck;
  txHashes: string[];
  error?: string | null;
  createdAt: string;
  executedAt?: string | null;
}

export interface CreateIntentRequest {
  prompt: string;
}

export interface TxResult {
  txHash: string;
  status: string;
  explorerUrl: string;
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  payment: 'payment',
  intent: 'intent',
  stream_created: 'stream_created',
  stream_claimed: 'stream_claimed',
  deposit: 'deposit',
  paused: 'paused',
  unpaused: 'unpaused',
  guardrail: 'guardrail',
  employee: 'employee',
} as const;

export interface ActivityEvent {
  type: ActivityEventType;
  title: string;
  subtitle: string;
  amount?: string | null;
  amountFormatted?: string | null;
  symbol?: string | null;
  address?: string | null;
  txHash?: string | null;
  explorerUrl?: string | null;
  blockNumber?: number | null;
  timestamp?: number | null;
}