# lib/api-zod/src/generated/types/appConfig.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Audd Flow — AI-governed smart treasury & automated payroll on BOT Chain
 * OpenAPI spec version: 0.1.0
 */
import type { DeploymentInfo } from './deploymentInfo';
import type { NetworkConfig } from './networkConfig';

export interface AppConfig {
  network: NetworkConfig;
  deployment: DeploymentInfo;
}