# lib/api-zod/src/generated/types/guardrail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Audd Flow — AI-governed smart treasury & automated payroll on BOT Chain
 * OpenAPI spec version: 0.1.0
 */

export interface Guardrail {
  token: string;
  symbol: string;
  perTxCap: string;
  perTxCapFormatted: string;
  dailyCap: string;
  dailyCapFormatted: string;
  spentToday: string;
  spentTodayFormatted: string;
  remainingToday: string;
  remainingTodayFormatted: string;
}