# lib/api-zod/src/generated/types/guardrailCheck.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Audd Flow — AI-governed smart treasury & automated payroll on BOT Chain
 * OpenAPI spec version: 0.1.0
 */

export interface GuardrailCheck {
  ok: boolean;
  violations: string[];
}