# lib/api-zod/src/generated/types/networkConfig.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Audd Flow — AI-governed smart treasury & automated payroll on BOT Chain
 * OpenAPI spec version: 0.1.0
 */
import type { TokenInfo } from './tokenInfo';

export interface NetworkConfig {
  name: string;
  chainId: number;
  rpcUrl: string;
  explorerUrl: string;
  faucetUrl: string;
  nativeToken: TokenInfo;
  ausdToken?: TokenInfo | null;
}